+++ src/data_value.rs
@@ -1,19 +1,9 @@
 use crate::{
-    ast::{
-        Expr,
-        Value,
-    },
+    ast::{Expr, Value},
     data_type::DataType,
 };
-use chrono::{
-    offset::FixedOffset,
-    DateTime,
-    Local,
-    NaiveDateTime,
-    Utc,
-};
+use chrono::{offset::FixedOffset, DateTime, Local, NaiveDateTime, Utc};
 use sql_ast::ast as sql;
-use std::net::IpAddr;
 use url::Url;
 use uuid::Uuid;
 
@@ -107,12 +97,10 @@ impl Into<u32> for DataValue {
             DataValue::I16(v) => v as u32,
             DataValue::I32(v) => v as u32,
             DataValue::I64(v) => v as u32,
-            _ => {
-                panic!(
-                    "unsupported conversion: {:?} to u32",
-                    self.get_data_type()
-                )
-            }
+            _ => panic!(
+                "unsupported conversion: {:?} to u32",
+                self.get_data_type()
+            ),
         }
     }
 }
@@ -128,12 +116,10 @@ impl Into<u64> for DataValue {
             DataValue::I16(v) => v as u64,
             DataValue::I32(v) => v as u64,
             DataValue::I64(v) => v as u64,
-            _ => {
-                panic!(
-                    "unsupported conversion: {:?} to u64",
-                    self.get_data_type()
-                )
-            }
+            _ => panic!(
+                "unsupported conversion: {:?} to u64",
+                self.get_data_type()
+            ),
         }
     }
 }
@@ -151,12 +137,10 @@ impl Into<f32> for DataValue {
             DataValue::I64(v) => v as f32,
             DataValue::F32(v) => v,
             DataValue::F64(v) => v as f32,
-            _ => {
-                panic!(
-                    "unsupported conversion: {:?} to f32",
-                    self.get_data_type()
-                )
-            }
+            _ => panic!(
+                "unsupported conversion: {:?} to f32",
+                self.get_data_type()
+            ),
         }
     }
 }
@@ -165,12 +149,10 @@ impl Into<String> for DataValue {
     fn into(self) -> String {
         match self {
             DataValue::Text(v) => v,
-            _ => {
-                panic!(
-                    "unsupported conversion: {:?} to String",
-                    self.get_data_type()
-                )
-            }
+            _ => panic!(
+                "unsupported conversion: {:?} to String",
+                self.get_data_type()
+            ),
         }
     }
 }
@@ -179,12 +161,10 @@ impl<'a> Into<&'a str> for &'a DataValue {
     fn into(self) -> &'a str {
         match self {
             DataValue::Text(ref v) => v,
-            _ => {
-                panic!(
-                    "unsupported conversion: {:?} to &str",
-                    self.get_data_type()
-                )
-            }
+            _ => panic!(
+                "unsupported conversion: {:?} to &str",
+                self.get_data_type()
+            ),
         }
     }
 }
@@ -229,57 +209,47 @@ pub fn cast_data_value(value: &Value, required_type: &DataType) -> DataValue {
         DataValue::Nil
     } else {
         match *value {
-            Value::Bool(v) => {
-                match *required_type {
-                    DataType::Bool => DataValue::Bool(v),
-                    DataType::U8 => DataValue::U8(if v { 1 } else { 0 }),
-                    DataType::U16 => DataValue::U16(if v { 1 } else { 0 }),
-                    DataType::U32 => DataValue::U32(if v { 1 } else { 0 }),
-                    DataType::U64 => DataValue::U64(if v { 1 } else { 0 }),
-                    DataType::I8 => DataValue::I8(if v { 1 } else { 0 }),
-                    DataType::I16 => DataValue::I16(if v { 1 } else { 0 }),
-                    DataType::I32 => DataValue::I32(if v { 1 } else { 0 }),
-                    DataType::I64 => DataValue::I64(if v { 1 } else { 0 }),
-                    _ => {
-                        panic!(
-                            "unsupported conversion from {:?} to {:?}",
-                            value, required_type
-                        )
-                    }
-                }
-            }
-            Value::Number(v) => {
-                match *required_type {
-                    DataType::U8 => DataValue::U8(v as u8),
-                    DataType::U16 => DataValue::U16(v as u16),
-                    DataType::U32 => DataValue::U32(v as u32),
-                    DataType::U64 => DataValue::U64(v as u64),
-                    DataType::I8 => DataValue::I8(v as i8),
-                    DataType::I16 => DataValue::I16(v as i16),
-                    DataType::I32 => DataValue::I32(v as i32),
-                    DataType::I64 => DataValue::I64(v as i64),
-                    DataType::F32 => DataValue::F32(v as f32),
-                    DataType::F64 => DataValue::F64(v as f64),
-                    _ => {
-                        panic!(
-                            "unsupported conversion from {:?} to {:?}",
-                            value, required_type
-                        )
-                    }
-                }
-            }
+            Value::Bool(v) => match *required_type {
+                DataType::Bool => DataValue::Bool(v),
+                DataType::U8 => DataValue::U8(if v { 1 } else { 0 }),
+                DataType::U16 => DataValue::U16(if v { 1 } else { 0 }),
+                DataType::U32 => DataValue::U32(if v { 1 } else { 0 }),
+                DataType::U64 => DataValue::U64(if v { 1 } else { 0 }),
+                DataType::I8 => DataValue::I8(if v { 1 } else { 0 }),
+                DataType::I16 => DataValue::I16(if v { 1 } else { 0 }),
+                DataType::I32 => DataValue::I32(if v { 1 } else { 0 }),
+                DataType::I64 => DataValue::I64(if v { 1 } else { 0 }),
+                _ => panic!(
+                    "unsupported conversion from {:?} to {:?}",
+                    value, required_type
+                ),
+            },
+            Value::Number(v) => match *required_type {
+                DataType::U8 => DataValue::U8(v as u8),
+                DataType::U16 => DataValue::U16(v as u16),
+                DataType::U32 => DataValue::U32(v as u32),
+                DataType::U64 => DataValue::U64(v as u64),
+                DataType::I8 => DataValue::I8(v as i8),
+                DataType::I16 => DataValue::I16(v as i16),
+                DataType::I32 => DataValue::I32(v as i32),
+                DataType::I64 => DataValue::I64(v as i64),
+                DataType::F32 => DataValue::F32(v as f32),
+                DataType::F64 => DataValue::F64(v as f64),
+                _ => panic!(
+                    "unsupported conversion from {:?} to {:?}",
+                    value, required_type
+                ),
+            },
             Value::String(ref v) => {
                 match *required_type {
                     DataType::Text => DataValue::Text(v.to_string()),
-                    DataType::Bool => {
-                        match v.as_ref() {
-                            "true" => DataValue::Bool(true),
-                            "false" => DataValue::Bool(false),
-                            "1" => DataValue::Bool(true),
-                            "0" => DataValue::Bool(false),
-                            _ => DataValue::Bool(false),
-                        }
-                    }
+                    DataType::Bool => match v.as_ref() {
+                        "true" => DataValue::Bool(true),
+                        "false" => DataValue::Bool(false),
+                        "1" => DataValue::Bool(true),
+                        "0" => DataValue::Bool(false),
+                        _ => DataValue::Bool(false),
+                    },
                     DataType::S8 => {
                         if let Ok(v) = v.parse::<u8>() {
                             DataValue::S8(v)
@@ -441,20 +411,16 @@ pub fn cast_data_value(value: &Value, required_type: &DataType) -> DataValue {
                     DataType::UuidSlug => DataValue::UuidSlug(v.to_string()),
                     //TODO: validate identifier
                     DataType::Ident => DataValue::Ident(v.to_string()),
-                    _ => {
-                        panic!(
-                            "unsupported conversion from {:?} to {:?}",
-                            value, required_type
-                        )
-                    }
+                    _ => panic!(
+                        "unsupported conversion from {:?} to {:?}",
+                        value, required_type
+                    ),
                 }
             }
-            _ => {
-                panic!(
-                    "unsupported conversion from {:?} to {:?}",
-                    value, required_type
-                )
-            }
+            _ => panic!(
+                "unsupported conversion from {:?} to {:?}",
+                value, required_type
+            ),
         }
     }
 }