+++ src/parser.rs
@@ -1,13 +1,5 @@
 use crate::ast::*;
-use pom::parser::{
-    call,
-    end,
-    is_a,
-    one_of,
-    sym,
-    tag,
-    Parser,
-};
+use pom::parser::{call, is_a, one_of, sym, tag, Parser};
 use std::iter::FromIterator;
 use utils::*;
 
@@ -46,7 +38,7 @@ where
                                 }
                             }
                         }
-                        Err(e) => {
+                        Err(_e) => {
                             // the separator does not match, just break
                             break;
                         }
@@ -307,20 +299,15 @@ pub fn filter_expr<'a>() -> Parser<'a, char, Expr> {
 fn from_table<'a>() -> Parser<'a, char, FromTable> {
     (table().expect("Expecting a valid table name")
         + (join_type() + call(from_table)))
-    .map(|(from, (join_type, from_table))| {
-        FromTable {
-            from,
-            join: Some((join_type, Box::new(from_table))),
-        }
+    .map(|(from, (join_type, from_table))| FromTable {
+        from,
+        join: Some((join_type, Box::new(from_table))),
     }) | (table().expect("Expecting a valid table name")
         + (join_type() + call(from_table)).opt())
-    .map(|(from, join)| {
-        FromTable {
-            from,
-            join: join.map(|(join_type, from_table)| {
-                (join_type, Box::new(from_table))
-            }),
-        }
+    .map(|(from, join)| FromTable {
+        from,
+        join: join
+            .map(|(join_type, from_table)| (join_type, Box::new(from_table))),
     })
 }